requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
Django==3.2.9
djangorestframework==3.12.4
google-api-core==2.2.1
google-auth==2.3.3
google-cloud-dialogflow==2.9.1
googleapis-common-protos==1.53.0
grpcio==1.41.1
grpcio-status==1.41.1
idna==3.3
packaging==21.2
proto-plus==1.19.7
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pytz==2021.3
requests==2.26.0
rsa==4.7.2
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
urllib3==1.26.7
